remove old and test components
