+++ src/app/core/components/dashboard/dashboard.component.ts
@@ -19,7 +19,6 @@ import { WidgetNumericComponent } from '../../../widgets/widget-numeric/widget-n
 import { WidgetDatetimeComponent } from '../../../widgets/widget-datetime/widget-datetime.component';
 import { WidgetBooleanSwitchComponent } from '../../../widgets/widget-boolean-switch/widget-boolean-switch.component';
 import { WidgetAutopilotComponent } from '../../../widgets/widget-autopilot/widget-autopilot.component';
-import { WidgetAutopilotV2Component } from '../../../widgets/widget-autopilot-v2/widget-autopilot-v2.component';
 import { WidgetDataChartComponent } from '../../../widgets/widget-data-chart/widget-data-chart.component';
 import { WidgetFreeboardskComponent } from '../../../widgets/widget-freeboardsk/widget-freeboardsk.component';
 import { WidgetGaugeNgCompassComponent } from '../../../widgets/widget-gauge-ng-compass/widget-gauge-ng-compass.component';
@@ -76,7 +75,6 @@ export class DashboardComponent implements AfterViewInit, OnDestroy{
       WidgetSteelGaugeComponent,
       WidgetFreeboardskComponent,
       WidgetAutopilotComponent,
-      WidgetAutopilotV2Component,
       WidgetDataChartComponent,
       WidgetRaceTimerComponent,
       WidgetIframeComponent,
