+++ src/app/widgets/svg-autopilot/svg-autopilot.component.scss
@@ -1,40 +1,58 @@
-.apScreen .apStencil {
-  position: absolute;
-  top: 0px;
-  left: 0px;
+:host {
+  display: block;
+  width: 100%;
+  font-family: Roboto;
+  overflow: hidden;
+  aspect-ratio: 2 / 1; // width:height = 2:1, so height is half the width
+  max-width: 100%;
+  max-height: 100%;
+}
+
+svg {
   width: 100%;
   height: 100%;
-  margin: auto;
-  border-radius: 5px;
-  z-index: 10;
-  vertical-align: middle;
-  background-color: var(--mat-sys-background);
-}
-
-.apScreen {
   display: block;
-  position: relative;
-  width: 360px;
-  height: calc(202px + 38.5px); // add status bar
-  margin: auto;
-  text-align: center;
-  border-radius: 5px;
-  border: 1px solid var(--kip-grey-dim-color);
-  background-color: var(--mat-sys-background);
+  overflow: hidden;
 }
 
-.compass-inner-ring {
-  stroke: var(--kip-contrast-color);
+.starboard  {
+  stroke: var(--kip-starboard-color);
+}
+
+.port {
+  stroke: var(--kip-port-color);
+}
+
+.contrast-text {
   fill: var(--kip-contrast-color);
+  stroke: var(--kip-contrast-color);
 }
 
-.compass-outer-ring {
-    stroke: var(--kip-grey-color);
+.contrast-text-dim {
+  fill: var(--kip-contrast-dim-color);
+  stroke: var(--kip-contrast-dim-color);
 }
 
-.apCompassOrnaments {
-  fill: var(--kip-kip-widget-card-background-color);
-  stroke: var(--mat-sys-background);
+.awa-text {
+  fill: var(--kip-orange-color);
+  stroke: var(--kip-orange-color);
+}
+
+.awa-text-dim {
+  fill: var(--kip-orange-dim-color);
+  stroke: var(--kip-orange-dim-color);
+}
+
+.card-color {
+  fill: var(--kip-widget-card-background-color);
+}
+
+.fixed-dial-background {
+  fill: var(--kip-grey-color);
+}
+
+.rotating-dial-Backplate {
+  fill: var(--mat-sys-background);
 }
 
 .rudderPort {
@@ -45,18 +63,7 @@
   fill: var(--kip-starboard-color);
 }
 
-.starboard  {
-  stop-color: var(--kip-starboard-color);
-}
-
-.port {
-  stop-color: var(--kip-port-color);
-}
-
-.app-wind {
-    fill: var(--kip-blue-color);
-}
-
-.apText {
-  fill: var(--kip-contrast-color);
+.boxes {
+  stroke: var(--kip-grey-color);
+  fill: var(--kip-grey-color);
 }
