--- src/app/widgets/widget-autopilot-v2/widget-autopilot-v2.component.scss
@@ -1,296 +1,0 @@
-@use '@angular/material' as mat;
-
-:host {
-  display: block;
-  height: 100%;
-  width: 100%;
-}
-
-.widget-host {
-  display: block;
-  height: 100%;
-  width: 100%;
-}
-
-.autopilot-grid-container {
-  box-sizing: border-box;
-  width: 100%;
-  height: 100%;
-  display: grid;
-  gap: 0.5rem 0.5rem;
-  padding: 10px;
-  grid-template:
-    "modes modes"   minmax(50px, 1fr)
-    "screen screen" minmax(50px, auto)
-    / 1fr 1fr;
-  grid-auto-rows: minmax(50px, 2fr);
-}
-
-.ap-modes {
-  grid-area: modes;
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: minmax(50px, 1fr);
-  gap: 0.5rem 0.5rem;
-}
-
-.ap-screen {
-  position: relative;
-  grid-area: screen;
-  width: 100%;
-  height: 100%;
-  max-height: 200px;
-  align-content: center;
-}
-
-.ap-screen-overlay {
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  width: 100%;
-  height: 100%;
-  z-index: 50;
-  vertical-align: middle;
-  text-align: center;
-  background-color: var(--mat-sys-background);
-}
-
-.ap-error-icon {
-  padding-top: 15px;
-}
-
-.ap-error-text {
-  color: var(--mat-sys-error-container);
-}
-
-.ap-btn {
-  width: 100%;
-  border-radius: 12px;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-blue-dim-color),
-    filled-container-height: 100%
-  ));
-}
-
-.ap-btn-ctrlbar {
-  padding: 4px;
-}
-
-.ap-btn-modes {
-  @include mat.button-overrides((
-    filled-horizontal-padding: 0rem
-  ));
-}
-
-.ap-btn-notifications {
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-yellow-dim-color),
-  ));
-}
-
-.ap-btn-engage {
-  @include mat.button-overrides((
-    filled-horizontal-padding: 0rem,
-    filled-container-color: var(--kip-zone-alarm-color),
-  ));
-}
-
-.svg-icon-ctrlbar {
-  width: 100%;
-  height: 100%;
-  margin: 0px;
-  padding: 0px;
-}
-
-.svg-icon {
-  width: 100%;
-  height: 100%;
-  margin: 0px;
-  padding: 0.5rem;
-}
-
-.svg-element {
-  width: 100%;
-  height: 100%;
-  padding: 0em;
-  text {
-    font-family: Roboto;
-    font-size: 25px;
-    font-weight: bold;
-    text-align: center;
-    text-anchor: middle;
-    font-style: normal;
-    font-variant: normal;
-    font-stretch: normal;
-    font-variant-ligatures: normal;
-    font-variant-caps: normal;
-    font-variant-numeric: normal;
-    font-variant-east-asian: normal;
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: none;
-  }
-  path {
-    fill: none;
-    fill-opacity: 0;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-    stroke-width :7.59058;
-    stroke-linecap: round;
-    stroke-miterlimit: 0;
-    stroke-dasharray: none;
-  }
-  path.tack {
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-  }
-}
-
-.svg-element-msg {
-  font-family: Roboto;
-  font-size: 12px;
-  font-weight: bold;
-  text-align: center;
-  text-anchor: middle;
-  font-style: normal;
-  font-variant: normal;
-  font-stretch: normal;
-  font-variant-ligatures: normal;
-  font-variant-caps: normal;
-  font-variant-numeric: normal;
-  font-variant-east-asian: normal;
-  fill: var(--kip-contrast-color);
-  fill-opacity: 1;
-  stroke: none;
-}
-
-.svg-element-note {
-  font-family: Roboto;
-  font-size: 0.5em;
-  font-weight: normal;
-  text-align: center;
-  text-anchor: middle;
-  font-style: italic;
-  font-variant: normal;
-  font-stretch: normal;
-  font-variant-ligatures: normal;
-  font-variant-caps: normal;
-  font-variant-numeric: normal;
-  font-variant-east-asian: normal;
-  fill: var(--kip-contrast-dim-color);
-  fill-opacity: 1;
-  stroke: none;
-}
-
-.svg-element-error-txt {
-  fill: var(--mat-sys-error-container);
-}
-
-.svg-element-notifications {
-  fill: var(--kip-contrast-color);
-}
-
-.svg-element-clear-notifications {
-  fill: var(--kip-contrast-color);
-  fill-opacity: 1;
-  stroke: none;
-}
-
-.svg-modes {
-  text {
-    font-size: 12px;
-  }
-}
-
-.svg-route {
-  text {
-    font-size: 25px;
-  }
-  path {
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-  }
-}
-
-.embed-widget {
-  border: 1px solid var(--mat-sys-surface-container-highest);
-  border-radius: 12px;
-}
-
-@media (max-width: 600px) {
-  .autopilot-grid-container {
-    padding: 0.5rem;
-    gap: 0.25rem;
-  }
-
-  .autopilot-btn {
-    font-size: 0.95rem;
-    padding: 0.4em 0.5em;
-  }
-}
-
-.widgetOverlay {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  right: 0;
-  bottom: 0;
-  color: transparent;
-  z-index: 21;
-}
-
-.svg-menu-overlay {
-  position: absolute;
-  top: 0; left: 0; right: 0; bottom: 0;
-  background: rgba(0,0,0,0.8);
-  z-index: 19;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.svg-menu-box {
-  fill: var(--kip-blue-color);
-  stroke: none;
-  border-radius: 12px;
-}
-
-.svg-menu {
-  max-width: 80%;
-  background: transparent;
-  border-radius: 12px;
-  box-shadow: 0 4px 24px rgba(0,0,0,0.15);
-  z-index: 20;
-  display: block;
-  font-family: Roboto;
-  font-size: 1.1rem;
-  font-weight: bold;
-  text-anchor: middle;
-  text-align: center;
-  alignment-baseline: middle;
-  text {
-    fill: var(--kip-contrast-color);
-  }
-}
-
-.svg-menu-item {
-  fill: transparent;
-  stroke: none;
-}
-
-.cancel {
-  fill: var(--kip-grey-color);
-  stroke: none;
-}
-
-.counterText {
-  font-weight: bold;
-  font-size: 7em;
-  margin-top: 0px;
-  margin-bottom: 0px;
-}
