+++ src/app/widgets/widget-autopilot/widget-autopilot.component.ts
@@ -1,23 +1,32 @@
-import { Component, OnInit, OnDestroy, viewChild, inject, effect } from '@angular/core';
-import { Subscription } from 'rxjs';
-import { MatButton, MatMiniFabButton } from '@angular/material/button';
-
-import { SignalkRequestsService, skRequest } from '../../core/services/signalk-requests.service';
+import { Component, OnInit, OnDestroy, viewChild, inject, signal, effect, untracked, DestroyRef } from '@angular/core';
+import { MatButton, MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { TitleCasePipe } from '@angular/common';
+import { MatBadgeModule } from '@angular/material/badge';
 import { BaseWidgetComponent } from '../../core/utils/base-widget.component';
 import { WidgetHostComponent } from '../../core/components/widget-host/widget-host.component';
-import { IWidgetSvcConfig } from '../../core/interfaces/widgets-interface';
-import { MatBadge } from '@angular/material/badge';
-import { NgIf } from '@angular/common';
+import { IWidget, IWidgetSvcConfig } from '../../core/interfaces/widgets-interface';
 import { SvgAutopilotComponent } from '../svg-autopilot/svg-autopilot.component';
+import { WidgetPositionComponent } from '../widget-position/widget-position.component';
+import { WidgetNumericComponent } from '../widget-numeric/widget-numeric.component';
+import { WidgetDatetimeComponent } from "../widget-datetime/widget-datetime.component";
+import { DashboardService } from '../../core/services/dashboard.service';
+import { SignalkRequestsService, skRequest } from '../../core/services/signalk-requests.service';
+import { isEqual } from 'lodash-es';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { SignalkPluginsService } from '../../core/services/signalk-plugins.service';
+import { AppService } from '../../core/services/app-service';
 
-const defaultPreferredDisplayMode = {
-  wind: 'windAngleApparent',
-  route: 'headingMag',
-  auto: 'headingMag',
-  standby: 'headingMag'
+interface CommandDefinition {
+  path: string;
+  value: string | number;
 }
 
-const commands = {
+interface CommandsMap {
+  [key: string]: CommandDefinition;
+}
+
+const commands: CommandsMap = {
   "auto":    {"path":"self.steering.autopilot.state","value":"auto"},
   "wind":    {"path":"self.steering.autopilot.state","value":"wind"},
   "route":   {"path":"self.steering.autopilot.state","value":"route"},
@@ -30,117 +39,325 @@ const commands = {
   "tackToStarboard":   {"path":"self.steering.autopilot.actions.tack","value":"starboard"},
   "advanceWaypoint":   {"path":"self.steering.autopilot.actions.advanceWaypoint","value":"1"}
 };
-
-const noData = '-- -- -- --';
-const noPilot = 'No pilot';
-const noHeading = '---&deg;';
 const countDownDefault: number = 5;
-const timeoutBlink = 250;
+
 
 @Component({
     selector: 'widget-autopilot',
     templateUrl: './widget-autopilot.component.html',
     styleUrls: ['./widget-autopilot.component.scss'],
     standalone: true,
-    imports: [WidgetHostComponent, MatButton, SvgAutopilotComponent, MatMiniFabButton, NgIf, MatBadge]
+    imports: [WidgetHostComponent, SvgAutopilotComponent, MatButtonModule, TitleCasePipe, MatIconModule, MatBadgeModule, WidgetPositionComponent, WidgetNumericComponent, WidgetDatetimeComponent],
 })
 export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnInit, OnDestroy {
-  signalkRequestsService = inject(SignalkRequestsService);
+  private readonly signalkRequestsService = inject(SignalkRequestsService);
+  protected readonly dashboard = inject(DashboardService);
+  private readonly _destroyRef = inject(DestroyRef);
+  private readonly _plugins = inject(SignalkPluginsService);
+  protected readonly _app = inject(AppService);
 
-  // AP keypad
-  readonly powerBtn = viewChild<MatButton>('powerBtn');
-  readonly stbTackBtn = viewChild<MatButton>('stbTackBtn');
-  readonly plus1Btn = viewChild<MatButton>('plus1Btn');
-  readonly minus1Btn = viewChild<MatButton>('minus1Btn');
-  readonly prtTackBtn = viewChild<MatButton>('prtTackBtn');
-  readonly standbyBtn = viewChild<MatButton>('standbyBtn');
-  readonly plus10Btn = viewChild<MatButton>('plus10Btn');
-  readonly minus10Btn = viewChild<MatButton>('minus10Btn');
-  readonly autoBtn = viewChild<MatButton>('autoBtn');
-  readonly windModeBtn = viewChild<MatButton>('windModeBtn');
-  readonly trackModeBtn = viewChild<MatButton>('trackModeBtn');
-  readonly muteBtn = viewChild<MatButton>('muteBtn');
-  readonly messageBtn = viewChild<MatButton>('messageBtn');
+  // Keypad buttons & layout
+  protected apGrid = signal('none'); // Autopilot button grid visibility
+  protected readonly plus1Btn = viewChild.required<MatButton>('plus1Btn');
+  protected readonly minus1Btn = viewChild.required<MatButton>('minus1Btn');
+  protected readonly plus10Btn = viewChild.required<MatButton>('plus10Btn');
+  protected readonly minus10Btn = viewChild.required<MatButton>('minus10Btn');
+  protected readonly stbTackBtn = viewChild.required<MatButton>('stbTackBtn');
+  protected readonly prtTackBtn = viewChild.required<MatButton>('prtTackBtn');
+  protected readonly modesBtn = viewChild.required<MatButton>('modesBtn');
+  protected readonly engageBtn = viewChild.required<MatButton>('disengageBtn');
+  protected readonly advWptBtn = viewChild.required<MatButton>('advWptBtn');
+  // protected readonly dodgeBtn = viewChild.required<MatButton>('dodgeBtn');
 
-  // AP Screen
-  readonly apScreen = viewChild<any>('appSvgAutopilot');
+  protected apState = signal<string | null>(null); // Current Pilot Mode - used for display, keyboard state and buildCommand function
+  protected autopilotTargetHeading: number = 0;
+  protected autopilotTargetWindHeading: number = 0;
+  protected courseTargetHeading: number = 0;
+  protected heading: number = 0;
+  protected crossTrackError: number = 0;
+  protected windAngleApparent: number = 0;
+  protected rudder: number = 0;
 
-  displayName: string;
+  // Widget messaging countdown
+  protected countdownOverlayVisibility = signal<string>('hidden');
+  protected countdownOverlayText = signal<string>('');
+  protected msgOverlayVisibility = signal<string>('hidden');
+  protected msgOverlayText = signal<string>('');
+  protected errorOverlayVisibility = signal<string>('hidden');
+  protected errorOverlayText = signal<string>('');
+  private handleCountDownCounterTimeout = null;
+  private handleConfirmActionTimeout = null;
+  private handleDisplayErrorTimeout = null;
+  protected countDownValue: number = -1;
+  private actionToBeConfirmed: string = "";
 
-  currentAPState: any = null;      // Current Pilot Mode - used for display, keyboard state and buildCommand function
-  currentAPTargetAppWind: number = 0;
-  currentHeading: number = 0;
-  currentAppWindAngle: number = null;
-  currentRudder: number = null;
+  // Mode Menu
+  protected menuOpen = signal<boolean>(false);
+  protected menuItems = [
+    { label: 'Cancel', action: 'cancel' }
+  ];
+  protected readonly itemHeight = 60;
+  protected readonly padding = 20;
 
-  skApNotificationSub =  new Subscription;
-  skRequestSub = new Subscription; // signalk-Request result observer
-
-  // Widget var
-  handleCountDownCounterTimeout = null;
-  handleConfirmActionTimeout = null;
-  handleMessageTimeout = null;
-  handleReceiveTimeout = null;
-  handleDisplayErrorTimeout = null;
-  countDownValue: number = 0;
-  actionToBeConfirmed: string = "";
-  skPathToAck: string = "";
-  isWChecked: boolean = false;       // used for Wind toggle
-  isTChecked: boolean = false;       // used for Track toggle
-  isApConnected: boolean = false;
-  notificationsArray = {};
-  alarmsCount: number = 0;
-
-  notificationTest = {};
+  // Sub widget properties
+  private embedWidgetColor = 'contrast';
+  protected nextWptProperties = signal<IWidget>({
+    type: "widget-position",
+    uuid: "db473695-42b1-4435-9d3d-ac2f27bf9665",
+    config: {
+      displayName: "Next WPT",
+      filterSelfPaths: true,
+      paths: {
+        "longPath": {
+          description: "Longitude",
+          path: "self.navigation.courseRhumbline.nextPoint.position.longitude",
+          source: "default",
+          pathType: "number",
+          isPathConfigurable: true,
+          convertUnitTo: "longitudeMin",
+          showPathSkUnitsFilter: true,
+          pathSkUnitsFilter: null,
+          sampleTime: 500
+        },
+        "latPath": {
+          description: "Latitude",
+          path: "self.navigation.courseGreatCircle.nextPoint.position.latitude",
+          source: "default",
+          pathType: "number",
+          isPathConfigurable: true,
+          convertUnitTo: "latitudeMin",
+          showPathSkUnitsFilter: true,
+          pathSkUnitsFilter: null,
+          sampleTime: 500
+        }
+      },
+      color: this.embedWidgetColor,
+      enableTimeout: false,
+      dataTimeout: 5
+    }
+  }, {equal: isEqual});
+  protected ttwProperties = signal<IWidget>({
+    type: "widget-numeric",
+    uuid: "ee022f2f-ee23-41a7-b0b1-0928dc28864d",
+    config: {
+      displayName: "TTWpt",
+      filterSelfPaths: true,
+      paths: {
+        numericPath: {
+          description: "Time To Waypoint",
+          path: "self.navigation.course.calcValues.timeToGo",
+          source: "default",
+          pathType: "number",
+          isPathConfigurable: true,
+          convertUnitTo: "HH:MM:SS",
+          showPathSkUnitsFilter: true,
+          pathSkUnitsFilter: null,
+          sampleTime: 500
+        }
+      },
+      showMax: false,
+      showMin: false,
+      numDecimal: 1,
+      numInt: 1,
+      color: this.embedWidgetColor,
+      enableTimeout: false,
+      dataTimeout: 5,
+      ignoreZones: false
+    }
+  }, {equal: isEqual});
+  protected etaProperties = signal<IWidget>({
+    type: "widget-datetime",
+    uuid: "544ca35d-18bf-4a90-9aa8-b12312e3fc60",
+    config: {
+      displayName: "ETA",
+      filterSelfPaths: true,
+      paths: {
+        gaugePath: {
+          description: "Estimated Time Of Arrival Date",
+          path: "self.navigation.course.calcValues.estimatedTimeOfArrival",
+          source: "default",
+          pathType: "Date",
+          isPathConfigurable: true,
+          sampleTime: 500
+        }
+      },
+      dateFormat: "EEE HH:mm",
+      dateTimezone: "System Timezone -",
+      color: this.embedWidgetColor,
+      enableTimeout: false,
+      dataTimeout: 5
+    }
+  }, {equal: isEqual});
+  protected dtwProperties = signal<IWidget>({
+    type : "widget-numeric",
+    uuid : "ee02ef2f-ee23-41a7-b0b1-0928dc28864d",
+    config : {
+      displayName  : "DTWpt",
+      filterSelfPaths  : true,
+      paths  : {
+        numericPath  : {
+          description  : "Distance To Waypoint",
+          path  : "self.navigation.course.calcValues.distance",
+          source  : "default",
+          pathType  : "number",
+          isPathConfigurable  : true,
+          convertUnitTo  : "nm",
+          showPathSkUnitsFilter  : true,
+          pathSkUnitsFilter  : null,
+          sampleTime  : 500
+        }
+      },
+      showMax  : false,
+      showMin  : false,
+      numDecimal  : 1,
+      numInt  : 1,
+      color  : this.embedWidgetColor,
+      enableTimeout  : false,
+      dataTimeout  : 5,
+      ignoreZones  : false
+    }
+  }, {equal: isEqual});
+  protected xteProperties = signal<IWidget>({
+    type: "widget-numeric",
+    uuid: "9234856b-7573-4154-a44f-3baf7c6f119c",
+    config: {
+      displayName: "XTE",
+      filterSelfPaths: true,
+      paths: {
+        numericPath: {
+          description: "Cross Track Error",
+          path: "self.navigation.course.calcValues.crossTrackError",
+          source: "default",
+          pathType: "number",
+          isPathConfigurable: true,
+          convertUnitTo: "m",
+          showPathSkUnitsFilter: true,
+          pathSkUnitsFilter: null,
+          sampleTime: 500
+        }
+      },
+      showMax: false,
+      showMin: false,
+      numDecimal: 1,
+      numInt: 1,
+      color: this.embedWidgetColor,
+      enableTimeout: false,
+      dataTimeout: 5,
+      ignoreZones: false
+    }
+  }, {equal: isEqual});
+  protected brgProperties = signal<IWidget>({
+    type : "widget-numeric",
+    uuid : "27a73083-3786-4399-847e-f08f32cb13bd",
+    config : {
+      displayName : "Locked BRG (Mag)",
+      filterSelfPaths : true,
+      paths : {
+        numericPath : {
+          description : "Numeric Data",
+          path : "self.navigation.course.calcValues.bearingMagnetic",
+          source : "default",
+          pathType : "number",
+          isPathConfigurable : false,
+          convertUnitTo : "deg",
+          showPathSkUnitsFilter : false,
+          pathSkUnitsFilter : null,
+          sampleTime : 500
+        }
+      },
+      showMax : false,
+      showMin : false,
+      numDecimal : 0,
+      numInt : 1,
+      color : this.embedWidgetColor,
+      enableTimeout : false,
+      dataTimeout : 5,
+      ignoreZones : false
+    }
+  }, {equal: isEqual});
 
   constructor() {
     super();
 
     this.defaultConfig = {
-      displayName: 'N2k Autopilot',
       filterSelfPaths: true,
       paths: {
-        "apState": {
+        "autopilotState": {
           description: "Autopilot State",
           path: 'self.steering.autopilot.state',
           source: 'default',
           pathType: "string",
-          isPathConfigurable: true,
+          isPathConfigurable: false,
           showPathSkUnitsFilter: false,
           convertUnitTo: "",
           sampleTime: 500
         },
-        "apTargetHeadingMag": {
-          description: "Autopilot Target Heading Mag",
+        "autopilotTargetHeading": {
+          description: "Autopilot Target Magnetic Heading",
           path: 'self.steering.autopilot.target.headingMagnetic',
           source: 'default',
           pathType: "number",
           convertUnitTo: "deg",
-          isPathConfigurable: true,
+          isPathConfigurable: false,
           showPathSkUnitsFilter: false,
           pathSkUnitsFilter: 'rad',
           sampleTime: 500
         },
-        "apTargetWindAngleApp": {
-          description: "Autopilot Target Wind Angle Apparent",
+        "autopilotTargetWindHeading": {
+          description: "Autopilot Target Apparent Wind Angle",
           path: 'self.steering.autopilot.target.windAngleApparent',
           source: 'default',
           pathType: "number",
           convertUnitTo: "deg",
-          isPathConfigurable: true,
+          isPathConfigurable: false,
           showPathSkUnitsFilter: false,
           pathSkUnitsFilter: 'rad',
           sampleTime: 500
         },
-        // "apNotifications": {
-        //   description: "Autopilot Notifications",
-        //   path: 'self.notifications.autopilot.*', //TODO(David): need to add support for .* path subscription paths in sk service and widget config modal
+        // "courseTargetHeadingTrue": {
+        //   description: "Course Bearing True",
+        //   path: 'self.navigation.course.calcValues.bearingTrue',
         //   source: 'default',
-        //   pathType: "string",
-        //   convertUnitTo: "",
+        //   pathType: "number",
+        //   convertUnitTo: "deg",
         //   isPathConfigurable: false,
+        //   showPathSkUnitsFilter: false,
+        //   pathSkUnitsFilter: 'rad',
         //   sampleTime: 500
         // },
+        // "courseTargetHeadingMag": {
+        //   description: "Course Bearing Magnetic",
+        //   path: 'self.navigation.course.calcValues.bearingMagnetic',
+        //   source: 'default',
+        //   pathType: "number",
+        //   convertUnitTo: "deg",
+        //   isPathConfigurable: false,
+        //   showPathSkUnitsFilter: false,
+        //   pathSkUnitsFilter: 'rad',
+        //   sampleTime: 500
+        // },
+        "courseXte": {
+          description: "Cross Track Error",
+          path: "self.navigation.course.calcValues.crossTrackError",
+          source: "default",
+          pathType: "number",
+          isPathConfigurable: true,
+          convertUnitTo: "m",
+          showPathSkUnitsFilter: true,
+          pathSkUnitsFilter: 'm',
+          sampleTime: 500
+        },
+        "rudderAngle": {
+          description: "Rudder Angle",
+          path: 'self.steering.rudderAngle',
+          source: 'default',
+          pathType: "number",
+          convertUnitTo: "deg",
+          isPathConfigurable: false,
+          showPathSkUnitsFilter: false,
+          pathSkUnitsFilter: 'rad',
+          sampleTime: 500
+        },
         "headingMag": {
           description: "Heading Magnetic",
           path: 'self.navigation.headingMagnetic',
@@ -184,247 +401,193 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
           showPathSkUnitsFilter: false,
           pathSkUnitsFilter: 'rad',
           sampleTime: 500
-        },
-        "rudderAngle": {
-          description: "Rudder Angle",
-          path: 'self.steering.rudderAngle',
-          source: 'default',
-          pathType: "number",
-          convertUnitTo: "deg",
-          isPathConfigurable: true,
-          showPathSkUnitsFilter: false,
-          pathSkUnitsFilter: 'rad',
-          sampleTime: 500
-        },
+        }
       },
-      usage: {
-        "headingMag": ['wind', 'route', 'auto', 'standby'],
-        "headingTrue": ['wind', 'route', 'auto', 'standby'],
-        "windAngleApparent": ['wind'],
-        "windAngleTrueWater": ['wind'],
-      },
-      typeVal: {
-        "headingMag": 'Mag',
-        "headingTrue": 'True',
-        "windAngleApparent": 'AWA',
-        "windAngleTrueWater": 'TWA',
-      },
-      barColor: 'accent',     // theme palette to select
-      autoStart: false,
       invertRudder: true,
+      headingDirectionTrue: false,
+      courseDirectionTrue: false,
       enableTimeout: false,
       dataTimeout: 5
     };
+
+    effect(() => {
+      const mode = this.apState();
+
+      untracked(() => {
+        switch (mode) {
+          case "standby":
+            this.modesBtn().disabled = false;
+            this.engageBtn().disabled = true;
+            this.plus1Btn().disabled = true;
+            this.plus10Btn().disabled = true;
+            this.minus1Btn().disabled = true;
+            this.minus10Btn().disabled = true;
+            this.prtTackBtn().disabled = true;
+            this.stbTackBtn().disabled = true;
+            this.advWptBtn().disabled = true;
+            // this.dodgeBtn().disabled = true;
+            this.menuItems = [
+              { label: 'Auto', action: 'auto' },
+              { label: 'Wind', action: 'wind' },
+              { label: 'Cancel', action: 'cancel' }
+            ];
+            break;
+          case "auto":
+            this.modesBtn().disabled = false;
+            this.engageBtn().disabled = false;
+            this.plus1Btn().disabled = false;
+            this.plus10Btn().disabled = false;
+            this.minus1Btn().disabled = false;
+            this.minus10Btn().disabled = false;
+            this.prtTackBtn().disabled = true;
+            this.stbTackBtn().disabled = true;
+            this.advWptBtn().disabled = true;
+            // this.dodgeBtn().disabled = true;
+            this.menuItems = [
+              { label: 'Wind', action: 'wind' },
+              { label: 'Route', action: 'route' },
+              { label: 'Cancel', action: 'cancel' }
+            ];
+            break;
+          case "wind":
+            this.modesBtn().disabled = false;
+            this.engageBtn().disabled = false;
+            this.plus1Btn().disabled = false;
+            this.plus10Btn().disabled = false;
+            this.minus1Btn().disabled = false;
+            this.minus10Btn().disabled = false;
+            this.prtTackBtn().disabled = false;
+            this.stbTackBtn().disabled = false;
+            this.advWptBtn().disabled = true;
+            // this.dodgeBtn().disabled = true;
+            this.menuItems = [
+              { label: 'Auto', action: 'auto' },
+              { label: 'Cancel', action: 'cancel' }
+            ];
+            break;
+          case "route":
+            this.modesBtn().disabled = false;
+            this.engageBtn().disabled = false;
+            this.plus1Btn().disabled = true;
+            this.plus10Btn().disabled = true;
+            this.minus1Btn().disabled = true;
+            this.minus10Btn().disabled = true;
+            this.prtTackBtn().disabled = true;
+            this.stbTackBtn().disabled = true;
+            this.advWptBtn().disabled = false;
+            // this.dodgeBtn().disabled = false;
+            this.menuItems = [
+              { label: 'Auto', action: 'auto' },
+              { label: 'Cancel', action: 'cancel' }
+            ];
+            break;
+          default:
+            this.modesBtn().disabled = true;
+            this.engageBtn().disabled = true;
+            this.plus1Btn().disabled = true;
+            this.plus10Btn().disabled = true;
+            this.minus1Btn().disabled = true;
+            this.minus10Btn().disabled = true;
+            this.prtTackBtn().disabled = true;
+            this.stbTackBtn().disabled = true;
+            this.advWptBtn().disabled = true;
+            // this.dodgeBtn().disabled = true;
+            this.menuItems = [
+              { label: 'Auto', action: 'auto' },
+              { label: 'Wind', action: 'wind' },
+              { label: 'Route', action: 'route' },
+              { label: 'Cancel', action: 'cancel' }
+            ];
+        }
+        this.apGrid.set(mode ? 'grid' : 'none');
+      });
+    });
   }
 
   ngOnInit() {
     this.validateConfig();
-    if (this.widgetProperties.config.autoStart) {
-      setTimeout(() => {this.startApHead();});
-    }
-    // this.demoMode(); // demo mode for troubleshooting
+    this.configureEmbedWidgets();
+    this.startWidget();
+  }
+
+  private configureEmbedWidgets(): void {
+    const brg = this.widgetProperties.config.courseDirectionTrue;
+    this.brgProperties.update(prev => ({
+      ...prev,
+      config: {
+        ...prev.config,
+        paths: {
+          ...prev.config.paths,
+          numericPath: {
+            ...prev.config.paths['numericPath'],
+            path: brg ? "self.navigation.course.calcValues.bearingTrue" : "self.navigation.course.calcValues.bearingMagnetic",
+          }
+        },
+        displayName: brg ? "Locked BRG (True)" : "Locked BRG (Mag)",
+      }
+    }));
   }
 
   protected startWidget(): void {
+    this.startAllSubscriptions();
   }
 
   protected updateConfig(config: IWidgetSvcConfig): void {
     this.widgetProperties.config = config;
+    this.configureEmbedWidgets();
+    this.startAllSubscriptions();
   }
 
-  demoMode() {
-    // this.setNotificationMessage('{"path":"notifications.autopilot.PilotWarningWindShift","value":{"state":"alarm","message":"Pilot Warning Wind Shift"}}');
-  }
-
-  ngOnDestroy() {
-    this.destroyDataStreams();
-    this.unsubscribeSKRequest();
-    console.log("Autopilot Subs Stopped");
-  }
-
-  startAllSubscriptions() {
-    this.observeDataStream('apState', newValue => {
-        this.currentAPState = newValue.data.value;
-        this.SetKeyboardMode(this.currentAPState);
-      }
-    );
-
-    this.observeDataStream('headingMag', newValue => {
-        if (newValue.data.value === null) {
-          this.currentHeading = 0;
-        } else {
-          this.currentHeading = newValue.data.value;
-        }
-      }
-    );
-
-    this.observeDataStream('windAngleApparent', newValue => {
-        if (newValue.data.value === null) {
-          this.currentAppWindAngle = null;
-          return;
-        }
-
-        if (newValue.data.value < 0) {// stb
-          this.currentAppWindAngle = 360 + newValue.data.value; // adding a negative number subtracts it...
-        } else {
-          this.currentAppWindAngle = newValue.data.value;
-        }
-      }
-    );
-
+  private startAllSubscriptions(): void {
+    this.unsubscribeDataStream();
+    this.observeDataStream('autopilotState', newValue => this.apState.set(newValue.data.value));
+    this.observeDataStream('autopilotTargetHeading', newValue => this.autopilotTargetHeading = newValue.data.value ? newValue.data.value : 0);
+    this.observeDataStream('autopilotTargetWindHeading', newValue => this.autopilotTargetWindHeading = newValue.data.value ? newValue.data.value : 0);
+    this.observeDataStream('courseXte', newValue => this.crossTrackError = newValue.data.value ? newValue.data.value : 0);
     this.observeDataStream('rudderAngle', newValue => {
         if (newValue.data.value === null) {
-          this.currentRudder = 0;
+          this.rudder = 0;
         } else {
-          this.currentRudder = this.widgetProperties.config.invertRudder ? -newValue.data.value : newValue.data.value;
+          this.rudder = this.widgetProperties.config.invertRudder ? -newValue.data.value : newValue.data.value;
         }
       }
     );
-
-    this.observeDataStream('apTargetWindAngleApp', newValue => {
-        if (newValue.data.value === null) {
-          this.currentAPTargetAppWind = 0;
-        } else {
-          this.currentAPTargetAppWind = newValue.data.value;
-        }
-      }
-    );
-
-    this.subscribeSKRequest();
-    // this.subscribeAPNotification();
-    console.log("Autopilot Subs Started");
+    // if (this.widgetProperties.config.courseDirectionTrue) {
+    //   this.observeDataStream('courseTargetHeadingTrue', newValue => this.courseTargetHeading = newValue.data.value ? newValue.data.value : 0);
+    // } else {
+    //   this.observeDataStream('courseTargetHeadingMag', newValue => this.courseTargetHeading = newValue.data.value ? newValue.data.value : 0);
+    // }
+    if (this.widgetProperties.config.headingDirectionTrue) {
+      this.observeDataStream('headingTrue', newValue => this.heading = newValue.data.value ? newValue.data.value : 0);
+    } else {
+      this.observeDataStream('headingMag', newValue => this.heading = newValue.data.value ? newValue.data.value : 0);
+    }
+    this.observeDataStream('windAngleApparent', newValue => this.windAngleApparent = newValue.data.value ? newValue.data.value : 0);
+    this.subscribePutResponse();
   }
 
-  stopAllSubscriptions() {
-    this.unsubscribeDataStream();
-    this.unsubscribeSKRequest();
-    // this.unsubscribeAPNotification();
-    console.log("Autopilot Subs Stopped");
-  }
-
-  subscribeSKRequest() {
-    this.skRequestSub = this.signalkRequestsService.subscribeRequest().subscribe(requestResult => {
+  private subscribePutResponse(): void {
+    this.signalkRequestsService.subscribeRequest().pipe(takeUntilDestroyed(this._destroyRef)).subscribe(requestResult => {
       if (requestResult.widgetUUID == this.widgetProperties.uuid) {
         this.commandReceived(requestResult);
       }
     });
   }
 
-  unsubscribeSKRequest() {
-    if (this.skRequestSub !== null) {
-      this.skRequestSub.unsubscribe();
-      this.skRequestSub = null;
+  protected toggleMenu(): void {
+    this.menuOpen.set(!this.menuOpen());
+  }
+
+  protected onMenuItemClick(action: string): void {
+    if (action === 'cancel') {
+      this.toggleMenu();
+      return;
     }
+    this.buildAndSendCommand(action);
+    this.menuOpen.set(false);
   }
 
-  addHeading(h1: number, h2: number) {
-    let h3 = h1 + h2;
-    while (h3 > 359) { h3 = h3 - 359; }
-    while (h3 < 0) { h3 = h3 + 359; }
-    return h3;
-  }
-
-  powerBtnClick(event: Event) {
-    if (!this.isApConnected) {
-      this.startApHead();
-    } else {
-      this.stopApHead();
-    }
-  }
-
-  startApHead() {
-    this.startAllSubscriptions();
-    this.widgetProperties.config.autoStart = true; // save power-on state to autostart or not
-    this.isApConnected = true;
-    this.muteBtn().disabled = true;
-    this.messageBtn().disabled = false;
-  }
-
-  stopApHead() {
-    this.muteBtn().disabled = true;
-    this.messageBtn().disabled = true;
-    this.windModeBtn().disabled = true;
-    this.trackModeBtn().disabled = true;
-    this.autoBtn().disabled = true;
-    this.standbyBtn().disabled = true;
-    this.plus1Btn().disabled = true;
-    this.plus10Btn().disabled = true;
-    this.minus1Btn().disabled = true;
-    this.minus10Btn().disabled = true;
-    this.prtTackBtn().disabled = true;
-    this.stbTackBtn().disabled = true;
-
-    this.apScreen().errorIconVisibility = 'hidden';
-
-    this.isApConnected = false; // hide ap screen
-    this.stopAllSubscriptions();
-    this.widgetProperties.config.autoStart = false; // save power on state to autostart or not
-  }
-
-  SetKeyboardMode(apMode: string) {
-    switch (apMode) {
-      case "standby":
-        this.trackModeBtn().disabled = true;
-        this.autoBtn().disabled = false;
-        this.standbyBtn().disabled = false;
-
-        this.windModeBtn().disabled = false;
-        this.plus1Btn().disabled = true;
-        this.plus10Btn().disabled = true;
-        this.minus1Btn().disabled = true;
-        this.minus10Btn().disabled = true;
-        this.prtTackBtn().disabled = true;
-        this.stbTackBtn().disabled = true;
-        break;
-
-      case "auto":
-        this.trackModeBtn().disabled = false;
-        this.autoBtn().disabled = false;
-        this.standbyBtn().disabled = false;
-
-        this.windModeBtn().disabled = false;
-        this.plus1Btn().disabled = false;
-        this.plus10Btn().disabled = false;
-        this.minus1Btn().disabled = false;
-        this.minus10Btn().disabled = false;
-        this.prtTackBtn().disabled = true;
-        this.stbTackBtn().disabled = true;
-        break;
-
-      case "wind":
-        this.trackModeBtn().disabled = true;
-        this.autoBtn().disabled = false;
-        this.standbyBtn().disabled = false;
-
-        this.windModeBtn().disabled = false;
-        this.plus1Btn().disabled = false;
-        this.plus10Btn().disabled = false;
-        this.minus1Btn().disabled = false;
-        this.minus10Btn().disabled = false;
-        this.prtTackBtn().disabled = false;
-        this.stbTackBtn().disabled = false;
-        break;
-
-      case "route":
-        this.trackModeBtn().disabled = false;
-        this.autoBtn().disabled = false;
-        this.standbyBtn().disabled = false;
-
-        this.windModeBtn().disabled = true;
-        this.plus1Btn().disabled = true;
-        this.plus10Btn().disabled = true;
-        this.minus1Btn().disabled = true;
-        this.minus10Btn().disabled = true;
-        this.prtTackBtn().disabled = true;
-        this.stbTackBtn().disabled = true;
-        break;
-
-      default:
-        break;
-    }
-  }
-
-  buildAndSendCommand(cmd: string) {
+  protected buildAndSendCommand(cmd: string) {
     let cmdAction = commands[cmd];
     if (typeof cmdAction === 'undefined') {
       alert('Unknown Autopilot command: ' + cmd);
@@ -433,20 +596,20 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
     if ((this.actionToBeConfirmed !== '') && (this.actionToBeConfirmed !== cmd)) {
       this.clearConfirmCmd();
     }
-    if (((cmd === 'tackToPort')||(cmd === 'tackToStarboard'))&&(this.actionToBeConfirmed === '')) {
+    if (((cmd === 'tackToPort') || (cmd === 'tackToStarboard')) && (this.actionToBeConfirmed === '')) {
       this.confirmTack(cmd);
       return null;
     }
-    if ((cmd === 'route')&&(this.currentAPState === 'route')&&(this.actionToBeConfirmed === '')) {
+    if ((cmd === 'route') && (this.apState() === 'route') && (this.actionToBeConfirmed === '')) {
       this.confirmAdvanceWaypoint(cmd);
       return null;
     }
     if (this.actionToBeConfirmed === cmd) {
       this.clearConfirmCmd();
-      if ((cmd === 'tackToPort')||(cmd === 'tackToStarboard')) {
+      if ((cmd === 'tackToPort') || (cmd === 'tackToStarboard')) {
         this.sendCommand(cmdAction);
       }
-      if ((cmd === 'route')&&(this.currentAPState === 'route')) {
+      if ((cmd === 'route') && (this.apState() === 'route')) {
         this.sendCommand(commands['advanceWaypoint']);
       }
       return null;
@@ -454,39 +617,34 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
     this.sendCommand(cmdAction);
   }
 
-  confirmAdvanceWaypoint(cmd: string) {
-    let message: string = "Repeat key <b>[Next Wpt]</b><br>to confirm<br>Advance Waypoint";
+  private confirmAdvanceWaypoint(cmd: string): void {
+    let message: string = "Repeat key [Adv Wpt] to confirm";
     this.startConfirmCmd(cmd, message);
   }
 
-  confirmTack(cmd: string) {
-    let message = "Repeat same key<br>to confirm<br>tack to ";
+  private confirmTack(cmd: string): void {
+    let direction: string = "";
     if (cmd === "tackToPort") {
-      message += "port";
+      direction = "Port";
       this.actionToBeConfirmed = cmd;
     } else if (cmd === "tackToStarboard") {
-        message += "starboard";
-        this.actionToBeConfirmed = cmd;
-      } else {
-          this.actionToBeConfirmed = "";
-          return null;
-        }
+      direction = "Starboard";
+      this.actionToBeConfirmed = cmd;
+    } else {
+      this.actionToBeConfirmed = "";
+      return null;
+    }
+
+    let message = `Repeat [Tack ${direction}] key to confirm`;
     this.startConfirmCmd(cmd, message);
   }
 
-  sendCommand(cmdAction) {
-    let requestId = this.signalkRequestsService.putRequest(cmdAction["path"], cmdAction["value"], this.widgetProperties.uuid);
-    this.apScreen().activityIconVisibility = "visible";
-    setTimeout(() => {this.apScreen().activityIconVisibility = 'hidden';}, timeoutBlink);
-
+  private sendCommand(cmdAction: CommandDefinition): void {
+    const requestId = this.signalkRequestsService.putRequest(cmdAction["path"], cmdAction["value"], this.widgetProperties.uuid);
     console.log("AP Action:\n" + JSON.stringify(cmdAction));
   }
 
-  commandReceived(cmdResult: skRequest) {
-    this.apScreen().activityIconVisibility = "visible";
-    clearTimeout(this.handleReceiveTimeout);
-    this.handleReceiveTimeout = setTimeout(() => {this.apScreen().activityIconVisibility = 'hidden';}, timeoutBlink);
-
+  private commandReceived(cmdResult: skRequest): void {
     if (cmdResult.statusCode != 200){
       this.displayApError(cmdResult);
     } else {
@@ -494,161 +652,64 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
     }
   }
 
-  startConfirmCmd(cmd: string, message: string) {
+  private startConfirmCmd(cmd: string, message: string): void {
     this.countDownValue = countDownDefault;
     this.actionToBeConfirmed = cmd;
 
-    const apScreen = this.apScreen();
-    apScreen.msgStencilInnerHTML = "<p>" + message + "</p>";
-    apScreen.msgStencilVisibility = "visible";
+    this.countdownOverlayText.set(message);
+    this.countdownOverlayVisibility.set("visible");
 
     this.updateCountDownCounter(message);
 
     clearTimeout(this.handleConfirmActionTimeout);
 
     this.handleConfirmActionTimeout = setTimeout(() => {
-      const apScreenValue = this.apScreen();
-      apScreenValue.msgStencilVisibility = "hidden";
-      apScreenValue.msgStencilInnerHTML = "";
+      this.countdownOverlayVisibility.set("hidden");
+      this.countdownOverlayText.set("");
       this.actionToBeConfirmed = "";
     }, 5000);
   }
 
-  clearConfirmCmd() : boolean {
+  private clearConfirmCmd(): void {
     clearTimeout(this.handleConfirmActionTimeout);
     clearTimeout(this.handleCountDownCounterTimeout);
     this.countDownValue = -1;
-    const apScreen = this.apScreen();
-    apScreen.msgStencilVisibility = "hidden";
-    apScreen.msgStencilInnerHTML = "";
+    this.countdownOverlayVisibility.set("hidden");
+    this.countdownOverlayText.set("");
     this.actionToBeConfirmed = '';
     return null;
   }
 
-  updateCountDownCounter(message: string) {
+  private updateCountDownCounter(message: string): void {
     if (this.countDownValue > 0) {
       clearTimeout(this.handleCountDownCounterTimeout);
-      this.apScreen().msgStencilInnerHTML = "<p>" + message + "</p>" + "<h1 class='counterText'>" + this.countDownValue.toString() + "</h1>";
+      this.countdownOverlayText.set(message);
       this.countDownValue -= 1;
       this.handleCountDownCounterTimeout = setTimeout(() => {
         this.updateCountDownCounter(message);
       }, 1000);
     } else {
+      this.countDownValue = -1;
         clearTimeout(this.handleCountDownCounterTimeout);
     }
   }
 
-  displayApError(cmdResult: skRequest) {
-    let errMsg = cmdResult.statusCode + " - " +cmdResult.statusCodeDescription;
+  private displayApError(cmdResult: skRequest): void {
+    let errMsg = cmdResult.statusCode + " - " + cmdResult.statusCodeDescription + ".";
     if (cmdResult.message){
-      errMsg = errMsg + " Server Message: " + cmdResult.message;
+      errMsg = errMsg + "Server Message: " + cmdResult.message + ".";
     }
-    const apScreen = this.apScreen();
-    apScreen.errorStencilInnerText = errMsg;
-    apScreen.errorStencilVisibility = "visible";
+    this.errorOverlayText.set(errMsg);
+    this.errorOverlayVisibility.set("visible");
 
     clearTimeout(this.handleDisplayErrorTimeout);
-
     this.handleDisplayErrorTimeout = setTimeout(() => {
-      const apScreenValue = this.apScreen();
-      apScreenValue.errorStencilVisibility = "hidden";
-      apScreenValue.errorStencilInnerText = "";
+      this.errorOverlayVisibility.set("hidden");
+      this.errorOverlayText.set("");
     }, 6000);
-    apScreen.errorIconVisibility = 'visible';
   }
 
-  getNextNotification(skPath: string): string {
-    let notificationsKeys = Object.keys(this.notificationsArray);
-    let newSkPathToAck: string = "";
-    let index: number = 0;
-    if (notificationsKeys.length > 0) {
-      if (typeof skPath !== "undefined") {
-        index = notificationsKeys.indexOf(skPath) + 1;
-      } else {
-          index = 0;
-        }
-      if (notificationsKeys.length <= index) {
-        index = 0;
-      }
-      newSkPathToAck = notificationsKeys[index];
-    }
-    return newSkPathToAck;
-  }
-
-  setNotificationMessage(value) {
-    const apScreen = this.apScreen();
-    apScreen.activityIconVisibility = "visible";
-    clearTimeout(this.handleReceiveTimeout);
-    this.handleReceiveTimeout = setTimeout(() => {this.apScreen().activityIconVisibility = 'hidden';}, timeoutBlink);
-
-    if (typeof value.path !== 'undefined') {
-      value.path = value.path.replace('notifications.', '');
-      if (typeof value.value !== 'undefined') {
-        if (value.value.state === 'normal') {
-          if (apScreen.messageInnerText === this.notificationsArray[value.path]) {
-            apScreen.messageInnerText = '';
-          }
-          delete this.notificationsArray[value.path]
-        } else {
-            this.notificationsArray[value.path] = value.value.message.replace("Pilot", "");
-            apScreen.messageInnerText = this.notificationsArray[value.path];
-          }
-      }
-    }
-    this.alarmsCount = Object.keys(this.notificationsArray).length;
-    if (this.alarmsCount > 0) {
-      this.muteBtn().disabled = false;
-      if (apScreen.messageInnerText == "") {
-        apScreen.messageInnerText = Object.keys(this.notificationsArray)[0];
-      }
-    } else {
-        this.muteBtn().disabled = true;
-        this.alarmsCount = 0;
-        apScreen.messageInnerText = "";
-      }
-  }
-
-  notificationToValue(skPathToAck: string): string {
-    let message: string = this.notificationsArray[skPathToAck];
-    if (typeof message == "undefined") {
-      message = "No alarm present...";
-    }
-    return message;
-  }
-
-  notificationScroll() {
-    if ((Object.keys(this.notificationsArray).length > 0) && (this.skPathToAck == "")) {
-      this.skPathToAck = Object.keys(this.notificationsArray)[0];
-    }
-
-    this.skPathToAck = this.getNextNotification(this.skPathToAck);
-    // Not sure about this DIV ??? May be message area???
-    // silenceScreenTextDiv.innerHTML = notificationToValue(skPathToAck);
-    const apScreen = this.apScreen();
-    apScreen.messageInnerText = this.notificationToValue(this.skPathToAck);
-    apScreen.messageVisibility = 'visible';
-    clearTimeout(this.handleMessageTimeout);
-    this.handleMessageTimeout = setTimeout(() => {
-      const apScreenValue = this.apScreen();
-      apScreenValue.messageInnerText = "";
-      apScreenValue.messageVisibility = 'hidden';
-    }, 2000);
-  }
-
-  sendSilence() {
-    const apScreen = this.apScreen();
-    if (apScreen.messageVisibility != 'visible') {
-      apScreen.messageVisibility = 'visible';
-
-      if ((Object.keys(this.notificationsArray).length > 0) && (this.skPathToAck == "")) {
-        this.skPathToAck = Object.keys(this.notificationsArray)[0];
-      }
-    } else {
-        if (this.skPathToAck !== "") {
-          this.sendCommand({"path":"notifications." + this.skPathToAck + ".state","value":"normal"});
-        }
-        apScreen.messageVisibility = 'hidden';
-      }
-      apScreen.messageInnerText = this.notificationToValue(this.skPathToAck);
+  ngOnDestroy() {
+   this.destroyDataStreams();
   }
 }
